refactor(listener): extract message router
